Added more UI stuff
Can see repeats, measures

and it jumps between measures nicely now

+++ src/groovy/com/expcork/musicconverter/music/Measure.groovy
@@ -14,7 +14,7 @@ class Measure {
 		def songStringMatcher = noteString =~ /([CDEFGABcdefgab][,\'']?)(\/?[0-9]*)?/
 		def notes = []
 		songStringMatcher.each {
-			int noteLength = NoteLengthUtil.modifyByAbcString(defaultNoteLength, it[2])
+			Double noteLength = NoteLengthUtil.modifyByAbcString(defaultNoteLength, it[2])
 			notes.add(Note.createFromString(it[1], noteLength))
 		}
